1. refactor code to support parallel CPU mining and organize modules in TREX pow library. 2. TRE-51 benchmark locally on parallel CPU mining and add the results to README.md.
