+++ consensus/trex-pow/Cargo.toml
@@ -36,5 +36,5 @@ sha2 = "0.10.2"
 criterion = "0.3"
 
 [[bench]]
-name = "pollard_rho_benchmark"
+name = "pollard_rho_hash"
 harness = false
